backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
supabase>=2.10.0
httpx>=0.28.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
celery[redis]>=5.4.0
redis>=5.2.0
playwright>=1.49.0
python-multipart>=0.0.18
anthropic>=0.40.0
openai>=1.57.0
google-auth>=2.36.0
google-auth-oauthlib>=1.2.1
google-api-python-client>=2.155.0
python-docx>=1.1.0
PyPDF2>=3.0.1
resend>=2.5.0
cryptography>=44.0.0
